added parameter sharing preset, pettingzoo env, minor fixes
