+++ all/presets/multiagent_atari/independent.py
@@ -4,7 +4,7 @@ from all.agents.multi.independent import IndependentMultiagent
 from all.logging import DummyWriter
 
 
-class IndependentMultiagentAtariPreset(Preset):
+class IndependentMultiagentPreset(Preset):
     def __init__(self, presets):
         self.presets = presets
